requirements.txt
Django==2.2.4
django-markdown==0.8.4
Markdown==3.1.1
pytz==2019.2
sqlparse==0.3.0
